[SPARKNLP-1174] Moving PDF parameters to HasPdfProperties
