+++ src/main/scala/com/johnsnowlabs/reader/util/HasPdfProperties.scala
@@ -16,6 +16,7 @@
 package com.johnsnowlabs.reader.util
 
 import com.johnsnowlabs.nlp.ParamsAndFeaturesWritable
+import com.johnsnowlabs.reader.util.pdf.TextStripperType
 import org.apache.spark.ml.param.{BooleanParam, IntParam, Param}
 
 trait HasPdfProperties extends ParamsAndFeaturesWritable {
@@ -26,6 +27,16 @@ trait HasPdfProperties extends ParamsAndFeaturesWritable {
   final val partitionNum = new IntParam(this, "partitionNum", "Number of partitions.")
   final val storeSplittedPdf =
     new BooleanParam(this, "storeSplittedPdf", "Force to store bytes content of splitted pdf.")
+  final val splitPage = new BooleanParam(
+    this,
+    "splitPage",
+    "Enable/disable splitting per page to identify page numbers and improve performance.")
+  final val onlyPageNum = new BooleanParam(this, "onlyPageNum", "Extract only page numbers.")
+  final val textStripper = new Param[String](
+    this,
+    "textStripper",
+    "Text stripper type used for output layout and formatting")
+  final val sort = new BooleanParam(this, "sort", "Enable/disable sorting content on the page.")
 
   /** @group setParam */
   def setPageNumCol(value: String): this.type = set(pageNumCol, value)
@@ -39,10 +50,27 @@ trait HasPdfProperties extends ParamsAndFeaturesWritable {
   /** @group setParam */
   def setStoreSplittedPdf(value: Boolean): this.type = set(storeSplittedPdf, value)
 
+  /** @group setParam */
+  def setSplitPage(value: Boolean): this.type = set(splitPage, value)
+
+  /** @group setParam */
+  def setOnlyPageNum(value: Boolean): this.type = set(onlyPageNum, value)
+
+  /** @group setParam */
+  def setTextStripper(value: String): this.type = set(textStripper, value)
+
+  /** @group setParam */
+  def setSort(value: Boolean): this.type = set(sort, value)
+
   setDefault(
     pageNumCol -> "pagenum",
     originCol -> "path",
     partitionNum -> 0,
-    storeSplittedPdf -> false)
+    storeSplittedPdf -> false,
+    onlyPageNum -> false,
+    splitPage -> true,
+    sort -> false,
+    textStripper -> TextStripperType.PDF_TEXT_STRIPPER
+  )
 
 }
